feat: refactor and performance changes
