+++ set-env.ts
@@ -26,7 +26,7 @@ const is_beta = release === "BETA";
 const is_canary = release === "CANARY";
 const is_dev = release === "DEV";
 
-const version = "2.9.9";
+const version = "2.9.10";
 
 // Configure Angular `environment.ts` file path
 const targetPath = "./src/environments/environment.ts";
