+++ src/app/components/authenticated-v2/components/changelog-dialog/changelog-dialog.component.html
@@ -24,7 +24,12 @@
     changes. Note that you can always look at the changelogs in the Help tab.
   </div>
 
-  <app-changelog-list></app-changelog-list>
+  <app-changelog-list
+    *ngIf="showChangelogList"
+    [itemsToShow]="changelogItemsToShow"></app-changelog-list>
+  <div *ngIf="!showChangelogList" style="display: flex; justify-content: center; padding: 20px">
+    Loading changelog...
+  </div>
 </mat-dialog-content>
 <mat-dialog-actions align="end">
   <button [mat-dialog-close]="true" cdkFocusInitial color="primary" mat-raised-button
