+++ src/app/components/authenticated-v2/components/changelog-dialog/changelog-dialog.component.ts
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-import { Component } from "@angular/core";
+import { Component, OnInit } from "@angular/core";
 import { ChangelogService } from "../../../../services/changelog.service";
 
 @Component({
@@ -23,6 +23,31 @@ import { ChangelogService } from "../../../../services/changelog.service";
   templateUrl: "./changelog-dialog.component.html",
   styleUrls: ["./changelog-dialog.component.css"],
 })
-export class ChangelogDialogComponent {
+export class ChangelogDialogComponent implements OnInit {
+  showChangelogList = false;
+  changelogItemsToShow = 1; // Initially show only the first item
+
   constructor(public changelog: ChangelogService) {}
+
+  ngOnInit() {
+    if (this.isIOS()) {
+      // Show first item immediately, then all items after delay
+      this.showChangelogList = true;
+      // Delay loading all changelog items by 10 seconds on iOS because WebKit is stinky and it crashes for no good reason
+      // (The reason appears to add too much junk from the manifest and the large DOM from the changelog list, but it's hard to be sure)
+      // I could diagnose it better if WebKit had any sort of decent debugging tools, but it doesn't, so this is the best workaround I can come up with
+      // Yes, I'm a bit angry :madcat:, I refactored too much shit, which was good, but unnecessary
+      setTimeout(() => {
+        this.changelogItemsToShow = -1; // Show all items
+      }, 10000);
+    } else {
+      // Show all items immediately on non-iOS devices
+      this.showChangelogList = true;
+      this.changelogItemsToShow = -1;
+    }
+  }
+
+  private isIOS(): boolean {
+    return /iPad|iPhone|iPod/.test(navigator.userAgent);
+  }
 }
