+++ src/app/components/authenticated-v2/overlays/gearset-tooltip/gearset-tooltip.component.html
@@ -23,6 +23,7 @@
           <img
             alt="{{ gearset.displayProperties.name }}"
             class="perk-icon"
+            loading="lazy"
             src="https://bungie.net/{{ gearset.displayProperties.icon }}" />
         </td>
       </tr>
