+++ src/app/components/authenticated-v2/overlays/stat-cooldown-tooltip/stat-cooldown-tooltip.component.html
@@ -38,7 +38,10 @@
           ">
           <ng-container *ngVar="valueFunction(entry, tier) as currentTierValue">
             <td>
-              <img src="https://www.bungie.net/{{ entry.icon }}" *ngIf="entry.icon" />
+              <img
+                src="https://www.bungie.net/{{ entry.icon }}"
+                loading="lazy"
+                *ngIf="entry.icon" />
               {{ entry.name }}
             </td>
             <td>{{ formatEntry(entry, currentTierValue) }}</td>
@@ -58,7 +61,7 @@
                 <td *ngIf="differenceTier !== tier">
                   <span [class.good]="differenceTier < tier" [class.bad]="differenceTier > tier">
                     ({{ tierDifference > 0 ? "+" : ""
-                    }}{{ tierDifference * 100 | number : "1.0-0" }}%)
+                    }}{{ tierDifference * 100 | number: "1.0-0" }}%)
                   </span>
                 </td>
               </ng-container>
