+++ src/app/components/authenticated-v2/settings/desired-mod-limit-selection/modslot-visualization/modslot-visualization.component.html
@@ -16,6 +16,7 @@
           <ng-container *ngSwitchCase="SlotType.SlotTypeGearset">
             <img
               [src]="'https://bungie.net/' + slot.icon"
+              loading="lazy"
               class="modslot-icon gearset"
               alt="Gearset"
               [matTooltip]="slot.name || ''" />
