+++ src/app/components/authenticated-v2/settings/desired-stat-selection/desired-stat-selection.component.ts
@@ -20,7 +20,7 @@ import { ArmorStat, ArmorStatNames, ARMORSTAT_ORDER } from "../../../../data/enu
 import { ConfigurationService } from "../../../../services/configuration.service";
 import { EnumDictionary } from "../../../../data/types/EnumDictionary";
 import { FixableSelection, getDefaultStatDict } from "../../../../data/buildConfiguration";
-import { UserInformationService } from "src/app/services/user-information.service";
+import { ArmorCalculatorService } from "../../../../services/armor-calculator.service";
 import { ModInformation } from "../../../../data/ModInformation";
 import { Subject } from "rxjs";
 import { takeUntil } from "rxjs/operators";
@@ -42,7 +42,7 @@ export class DesiredStatSelectionComponent implements OnInit, OnDestroy {
 
   constructor(
     public config: ConfigurationService,
-    private inventory: UserInformationService
+    private armorCalculator: ArmorCalculatorService
   ) {
     this.stats = ARMORSTAT_ORDER.map((value) => {
       return { name: (ArmorStatNames as any)[+value], value: +value };
@@ -64,7 +64,7 @@ export class DesiredStatSelectionComponent implements OnInit, OnDestroy {
       this.config_reduce_waste = c.tryLimitWastedStats;
     });
 
-    this.inventory.reachableTiers.pipe(takeUntil(this.ngUnsubscribe)).subscribe((d) => {
+    this.armorCalculator.reachableTiers.pipe(takeUntil(this.ngUnsubscribe)).subscribe((d) => {
       // Do not update if we get 0 results
       const tiers = d || [20, 20, 20, 20, 20, 20];
       if (tiers.filter((d) => d == 0).length < 6) {
@@ -72,7 +72,7 @@ export class DesiredStatSelectionComponent implements OnInit, OnDestroy {
       }
     });
 
-    this.inventory.armorResults.pipe(takeUntil(this.ngUnsubscribe)).subscribe((d) => {
+    this.armorCalculator.armorResults.pipe(takeUntil(this.ngUnsubscribe)).subscribe((d) => {
       // Do not update if we get 0 results
       const tiers = d.maximumPossibleTiers || [20, 20, 20, 20, 20, 20];
       if (tiers.filter((d) => d == 0).length < 6) {
