+++ src/app/components/authenticated-v2/subpages/account-config-page/account-config-page.component.ts
@@ -56,7 +56,7 @@ export class AccountConfigPageComponent {
 
   async resetDatabase() {
     await this.db.resetDatabase();
-    await this.inv.refreshManifestAndArmor(true, true);
+    await this.inv.refreshManifestAndInventory(true, true);
   }
 
   async resetEverything() {
